feat: add table
