+++ src/assets/styles/global/variables/_z-index.scss
@@ -1,0 +1,3 @@
+$z-index: (
+  'snackbar': 9999
+);
