+++ src/assets/styles/global/variables/index.scss
@@ -1,0 +1,1 @@
+@import './z-index';
