+++ src/ts/type/index.ts
@@ -4,4 +4,4 @@ export * from './Jwt'
 export * from './User'
 export * from './Enums'
 export * from './Snackbar'
-export * from './Apis'
+export * from './Api'
